fix get-block-styles by extracting overridden fns
