+++ packages/sdks/overrides/solid/src/functions/convert-style-object.ts
@@ -1,0 +1,13 @@
+const camelCaseToDashCase = (str = '') =>
+  str.replace(/[A-Z]/g, (match) => '-' + match.toLowerCase());
+
+export const convertStyleObject = (obj: Partial<CSSStyleDeclaration>) => {
+  if (!obj) {
+    return obj;
+  }
+  const newObj = {};
+  for (const key in obj) {
+    newObj[camelCaseToDashCase(key)] = obj[key];
+  }
+  return newObj;
+};
