+++ packages/sdks/src/functions/convert-style-object.ts
@@ -1,0 +1,3 @@
+export const convertStyleObject = (obj?: Partial<CSSStyleDeclaration>) => {
+  return obj;
+};
