+++ packages/sdks/src/functions/sanitize-styles.ts
@@ -1,0 +1,7 @@
+/**
+ * No-op function, sanitize on a per-framework basis by overriding.
+ */
+export const sanitizeBlockStyles = (
+  _styles: Partial<CSSStyleDeclaration>
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+): void => {};
